% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{align_DNA_fasta}
\alias{align_DNA_fasta}
\title{Align an input fasta file with Muscle}
\usage{
align_DNA_fasta(input_path, output_path, ...)
}
\arguments{
\item{input_path}{A path to a fasta file you want to import and align with MUSCLE.}

\item{output_path}{A file path to where you want to save the aligned fasta file.
If omitted a \code{DNAStringSet} is returned.}

\item{...}{other parameters passed to \code{msa::msaMuscle()}.}
}
\value{
A \code{DNAStringSet}. If \code{output_path} is specified it writes to file.
}
\description{
This function is a wrapper to circumvent 2 annoying issues that
the \href{http://dx.doi.org/10.1093/bioinformatics/btv494}{\code{msa}} package has
when aligning a fasta sequence, namely not being able to sort the output
fasta gapped file as the input sequences and not being able to write to fasta file.
}
\details{
This function only aligns DNA fasta files using an \code{R} implementation of the muscle
multiple sequences aligner. \href{https://www.drive5.com/muscle/}{MUSCLE}
is claimed to achieve both better average accuracy and
better speed than ClustalW2 or T-Coffee, depending on the chosen options
(as stated \href{https://www.ebi.ac.uk/Tools/msa/muscle/}{here}).
}
\examples{
# Input file
dna_fasta_path <- file.path(seqs_dir, "DNA/All_Seq.fasta")
# Output
dna_alg_path <- file.path(seqs_dir, "Alignments/All_Seq_Ordered_as_Input.afa")

align_DNA_fasta(input_path = dna_fasta_path, output_path = dna_alg_path)
}
