% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{check_fasta}
\alias{check_fasta}
\title{Helper function to check a fasta file that is going to be imported into \code{R}.}
\usage{
check_fasta(path)
}
\arguments{
\item{path}{A character indicating a \code{File/path/to/a/fasta/file.fasta}}
}
\value{
Nothing. It returns an error message only if the file doesn't exist or is empty.
}
\description{
Helper function to check a fasta file that is going to be imported into \code{R}.
}
