% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_files_utils.R
\name{check_file}
\alias{check_file}
\title{Helper function to check a file that is going to be imported into \code{R}.}
\usage{
check_file(path)
}
\arguments{
\item{path}{A character indicating a \code{File/path/to/a/file}}
}
\value{
Nothing. It returns an error message only if the file doesn't exist or is empty.
}
\description{
Helper function to check a file that is going to be imported into \code{R}.
}
