% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{col_freq}
\alias{col_freq}
\title{Helper function to use as a tidyverse verb}
\usage{
col_freq(column, alphabet, verbose = FALSE)
}
\arguments{
\item{column}{}

\item{alphabet}{}

\item{verbose}{Logical. Whether or not to print warning messages about case-sensitivity. Default F}
}
\value{
a column frequency
}
\description{
Helper function to use as a tidyverse verb
}
