% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{col_JS_divergence}
\alias{col_JS_divergence}
\title{Helper function to calculate the Jensen-Shannon divergence}
\usage{
col_JS_divergence(p, q)
}
\arguments{
\item{p}{A data.frame of positional probabilities}

\item{q}{Another data.frame of positional probabilities}
}
\value{
A numeric list
}
\description{
The 2 data.frames \code{p} and \code{q} must have the same dimensions. The calculated data.frame \code{m} is the mean between \code{p} and \code{q}
}
\examples{
H_i <- mapply(FUN = col_JS_divergence, p = ppm_p, q = ppm_q) 
}
