% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{col_sub_abs}
\alias{col_sub_abs}
\title{Helper function to calculate the sum of the absolute difference between 2 position probability matrices}
\usage{
col_sub_abs(p, q)
}
\arguments{
\item{p}{}

\item{q}{}
}
\value{
a numeric vector
}
\description{
Helper function to calculate the sum of the absolute difference between 2 position probability matrices
}
\examples{
sum_abs_delta_prob_l <- mapply(FUN = col_sub_abs, p = ppm_p, q = ppm_q) 
}
