% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils_plot.R
\name{count_snps_in_bins}
\alias{count_snps_in_bins}
\title{Get how many SNPs variation fall within each bin in the different regions of the bed file.}
\usage{
count_snps_in_bins(
  df,
  snp_ID_col = "RefSNP_ID",
  left_slop,
  right_slop,
  left_bins_num = 50,
  centre_bins_num = 50,
  right_bins_num = 25,
  start_centre_len = 39,
  end_centre_len = 39,
  check_bins = T
)
}
\arguments{
\item{df}{}

\item{snp_ID_col}{}

\item{left_slop}{}

\item{right_slop}{}

\item{left_bins_num}{}

\item{centre_bins_num}{number of bins to divide the centre region (before slopping).}

\item{right_bins_num}{}

\item{start_centre_len}{Do not bin all centre region at once. Just consider the first N part of the centre region.}

\item{end_centre_len}{Just consider the last N part of the centre region.}

\item{check_bins}{}
}
\value{
A data.frame
}
\description{
Get how many SNPs variation fall within each bin in the different regions of the bed file.
}
