% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{dds2counts}
\alias{dds2counts}
\title{Convert a DESeq2 dds object into a tibble}
\usage{
dds2counts(deseq_dataset, tidy = T, counts_are_genes = T, norm_counts = T)
}
\arguments{
\item{deseq_dataset}{A DESeq2 object.}

\item{tidy}{Logical, whether or not to pivot the data into a long format}

\item{counts_are_genes}{Logical, by default this function assumes the counts are from gene expression experiments where the IDs are ensembl gene IDs.}
}
\value{
A tibble
}
\description{
Convert a DESeq2 dds object into a tibble
}
\examples{
tidy_counts <- dds2counts(dds)
}
