% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2bits}
\alias{df2bits}
\title{Calculate the information content expressed in bits of sequences stored in a data.frame}
\usage{
df2bits(
  data,
  ID_col,
  alphabet,
  small_n_correction = FALSE,
  long_format = FALSE,
  ignore_case = FALSE
)
}
\arguments{
\item{data}{A data.frame with a minimum of 2 columns. One named \code{Sequence}, the other named as you prefer that will be specified with \code{ID_col}.}

\item{ID_col}{The name of the column in \code{data} to be used as the identifier of the \code{Sequence} column.}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default.}

\item{small_n_correction}{Apply a small correction to the Shannon Entropy. See details. Default \code{FALSE}.}

\item{long_format}{Logical. If \code{TRUE} reshape the bits into a tidy long data.frame format. Default \code{FALSE}.}

\item{ignore_case}{Logical. If \code{TRUE} the length of the alphabet is calculated ignoring the case of the alphabet.
Meaning that the maximum bits height will calculated on the case-insensitive length of the alphabet. See notes for more explanation. Default \code{FALSE}.}
}
\value{
A data.frame or a tidy long format data.frame
}
\description{
This function calculates the information content expressed in bits using the Shannon entropy.
Check the details for full explanation and formulas.
However, currently there's no support for LaTeX syntax for subscript text and fractions.
To display them properly once could copy-paste the details section in \href{https://www.overleaf.com/}{Overleaf}.
}
\details{
Given an alphabet of letters of length \emph{W} where every letter
defined as \emph{l} for which \emph{l} belongs to \emph{W},
we can represent the DNA alphabet as \eqn{l'} belongs to \emph{A,C,G,T} where \emph{W = 4}.
With a multiple sequence alignment of \eqn{N} sequences of length \emph{I} we denote
the information content expressed in bits of the letter \emph{l} at position \emph{i}
with \eqn{bits_l_,_i} we define the following
formula \deqn{bits_l_,_i = R(l,i) \times ( log_2(W) - (H_i + \epsilon) )}
where \eqn{H_i} is the Shannon entropy representing the uncertainty of
position \emph{i} is defined as:
\deqn{ -\sum_{i = 1}^{W} { p_l_i \times log_2 p_l_,_i }} where \eqn{p_l_i} is the
relative frequency (a.k.a. probability) of letter \eqn{l} at position \eqn{i};
\eqn{\epsilon} is the approximation for small-sample corrections,
i.e. a correction for an alignment of \emph{N} sequences in the alignment defined as
\deqn{\epsilon = \frac{1}{log_e{2}} \times \frac{W-1}{2N}} and \eqn{R(l,i)} sequences
position probability matrix containing the \eqn{p_l_i} for \emph{N} sequences.
}
\note{
When having an upper and lower case DNA sequence, with an \code{alphabet} that
as both 'ATGC' and 'atgc' one case force the maximum information
content to \code{log2(4)} instead of \code{log2(8)} by doing \code{ignore_case = TRUE}.
}
\examples{
df2bits(data, ID_col = 'Species', 
        alphabet = c('a', 'c', 'g', 't'), 
        small_n_correction = F, 
        long_format = T)
}
