% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2JSD}
\alias{df2JSD}
\title{Calculate Jensen-Shannon divergence (JSD) between 2 set of sequences.}
\usage{
df2JSD(df1, df2, ID_col, alphabet, long_format = FALSE)
}
\arguments{
\item{df1}{}

\item{df2}{}

\item{ID_col}{}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default.}
}
\value{
A data.frame or a tidy long formatted data.frame
}
\description{
Letters in the alphabet that are less abundant in \code{df2} sequences relative to \code{df1} sequences have positive values.
More abundant letters in \code{df2} sequences relative to \code{df1} have nevative values.
}
\examples{
df2JSD(df1 = data_p, df2 = data_q, ID_col, long_format = F)
}
