% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2mat}
\alias{df2mat}
\title{Turn a dataframe into a matrix}
\usage{
df2mat(data, ID_col)
}
\arguments{
\item{data}{}

\item{ID_col}{}
}
\value{
a matrix
}
\description{
Turn a dataframe into a matrix
}
