% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2pfm}
\alias{df2pfm}
\title{Turn a data.frame with sequences into a positional frequency matrix.}
\usage{
df2pfm(data, ID_col, alphabet)
}
\arguments{
\item{data}{A data.frame with a minimum of 2 columns. One named \code{Sequence}, the other named as you prefer that will be specified with \code{ID_col}.}

\item{ID_col}{The name of the column in \code{data} to be used as the identifier of the \code{Sequence} column.}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default.}
}
\value{
A data frame with positional frequencies.
}
\description{
Given a dataframe with a column called 'Sequence' it transforms the sequences into a matrix where the frequency of each letter in the alphabet is counted at every position
}
\details{
For a great song click \href{https://youtu.be/tqvvaY2LvuI?t=110}{here}.
}
