% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2ppm}
\alias{df2ppm}
\title{Turn a data.frame with sequences into a positional probability matrix.}
\usage{
df2ppm(data, ID_col, alphabet, c = 5)
}
\arguments{
\item{data}{A data.frame with a minimum of 2 columns. One named \code{Sequence}, the other named as you prefer that will be specified with \code{ID_col}.}

\item{ID_col}{The name of the column in \code{data} to be used as the identifier of the \code{Sequence} column.}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default.}

\item{c}{an integer used to round up the numbers}
}
\value{
A data frame with positional probabilities.
}
\description{
Generates a positional probability matrix (PPM).
}
\examples{
df2ppm(data, ID_col = 'Species', alphabet = c("A", "C", "G", "T")) 
}
