% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{df2pwm}
\alias{df2pwm}
\title{Create a PWM from a dataframe with sequences}
\usage{
df2pwm(
  data,
  ID_col,
  alphabet,
  bg_prob_letter,
  pseudocount_letter,
  long_format = FALSE
)
}
\arguments{
\item{data}{A data.frame with a minimum of 2 columns. One named \code{Sequence}, the other named as you prefer that will be specified with \code{ID_col}.}

\item{ID_col}{The name of the column in \code{data} to be used as the identifier of the \code{Sequence} column.}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default.}

\item{bg_prob_letter}{Explain well in details.}

\item{pseudocount_letter}{Explain well in details.}

\item{long_format}{Logical. If \code{TRUE} reshape the PWM into a tidy long data.frame format. Default \code{FALSE}.}
}
\value{
A data.frame or a tidy long format data.frame
}
\description{
Generate a positional weight matrix (PWM) from a set of sequences. This function comes with batteries included.
}
\details{
Here I'll explain more things, ideally with formulas if LaTeX syntax was supported.
}
\examples{
df2pwm(data, ID_col = 'Species', alphabet = c('a', 'c', 'g', 't'), long_format = T) 
}
