% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{ensembl_id_2_entrez}
\alias{ensembl_id_2_entrez}
\title{Convenient converter from ensembl ID to entrez ID}
\usage{
ensembl_id_2_entrez(
  ensembl_gene_id,
  only_entrez = TRUE,
  verbose = FALSE,
  mRt_objct = ensembl
)
}
\arguments{
\item{ensembl_gene_id}{A character specifying the ENSEMBL gene ID}

\item{only_entrez}{Logical, whether to return only the gene name or a \code{data.frame} with more info. Default \code{TRUE}}

\item{verbose}{Logical. Do you want me to be chatty?}

\item{mRt_objct}{A \code{biomaRt} object created with the \code{gimme_mart()} function.}
}
\description{
Convenient converter from ensembl ID to entrez ID
}
\examples{
# Example with human ensembl
ensembl <- gimme_mart()
ensembl_id_2_gene_name("ENSG00000000003")
"7105"
# Example with mouse ensembl
ensembl <- gimme_mart(species = "mmusculus", latest = T, use_mirror = F, 
                      out_dir = F, verbose = T)
ensembl_id_2_entrez("ENSMUSG00000067377")
"56496"
}
