% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{ensembl_id_2_gene_name}
\alias{ensembl_id_2_gene_name}
\title{Convenient converter from ensembl ID to gene name}
\usage{
ensembl_id_2_gene_name(
  ensembl_gene_id,
  only_gene_name = TRUE,
  verbose = FALSE,
  mRt_objct = ensembl
)
}
\arguments{
\item{ensembl_gene_id}{A character specifying the ENSEMBL gene ID}

\item{only_gene_name}{Logical, whether to return only the gene name or a \code{data.frame} with more info. Default \code{TRUE}}

\item{verbose}{Logical. Do you want me to be chatty?}

\item{mRt_objct}{A \code{biomaRt} object created with the \code{gimme_mart()} function.}
}
\description{
Convenient converter from ensembl ID to gene name
}
\examples{
ensembl <- gimme_mart()
ensembl_id_2_gene_name("ENSG00000000003")
"TSPAN6"
# Example with mouse ensembl
ensembl <- gimme_mart(species = "mmusculus", latest = T, use_mirror = F, 
                      out_dir = F, verbose = T)
ensembl_id_2_gene_name("ENSMUSG00000067377", only_gene_name = F)
     ensembl_gene_id external_gene_name chromosome_name start_position   gene_biotype
1 ENSMUSG00000067377             Tspan6               X      132791817 protein_coding

}
