% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{fasta2df}
\alias{fasta2df}
\title{Handy wrapper to import fasta files into a data.frame}
\usage{
fasta2df(path, ID_col = "fasta_header", ...)
}
\arguments{
\item{path}{A character indicating a \code{File/path/to/a/fasta/file.fasta}}

\item{ID_col}{A character specifying a column name for the info contained in the fasta header. Default is \code{'fasta_header'}.}

\item{...}{Additional parameters passed to \code{seqinr::read.fasta} like \code{seqtype = "DNA"}}
}
\value{
A data.frame
}
\description{
Handy wrapper to import fasta files into a data.frame
}
\examples{
df1 <- fasta2df(path = "/file/to/fasta/species.fasta", ID_col = "Species")
}
