% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{fasta2pim}
\alias{fasta2pim}
\title{Align a fasta file with MUSCLE and calculate the percentage of sequence identity of the aligned sequence}
\usage{
fasta2pim(input_path, percentage_identity = TRUE, ...)
}
\arguments{
\item{input_path}{A path to a DNA fasta file for which you want the pairwise sequence identity}

\item{percentage_identity}{Logical, return the percentage of sequence identity.
Default \code{TRUE}. If \code{FALSE} the squared root of the pairwise identity is returned. See details.}

\item{...}{Other parameters passed to \code{msa::msaMuscle()}.}
}
\value{
A matrix with aligned DNA sequence identity
}
\description{
This function first aligns a DNA fasta file with multiple sequences
using MUSCLE using \code{msa::msaMuscle()}, where you can pass additional parameters thanks to \code{...}.
Then the pairwise sequence indentity is calculated using \code{seqinr::dist.alignment()}.
Gaps in the alignment will be counted in the identity measure. The output
matrix is always ordered as the input fasta sequences.
}
\details{
By default this function returns the percentage of sequence identity
from 0 to 100. By setting \code{percentage_identity = FALSE}, the sqrt(1 - identity)
is returned. So if the identity between 2 sequences is 19\% the squared root
of (1.0 - 0.19) i.e. 0.9.
}
\note{
If you want to inspect the multiple sequence alignment used to calculate
the percentage of sequence identity you can write it to a fasta file
with \code{align_DNA_fasta()} using the same MUSCLE parameters.
}
\examples{
pim <- fasta2pim(input_path = dna_fasta_path)
}
