% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{filter_dds}
\alias{filter_dds}
\title{Filter lowly abundant genes with the possibility of doing a normalisation on the genes to correct for trended biases.
This function will not estimate size factors of a DESeq2 object, but it will add normalization factors
to the DESeq2 object if the cyclic_loess is TRUE.
Plotting options allow to see how the read distribution changes using different filtering and normalisation parameters}
\usage{
filter_dds(
  deseq_dataset,
  filt_method = c("sum", "mean", "max"),
  min_counts = 5,
  verbose = T,
  show_mltdnsty = T,
  cyclic_loess = F
)
}
\arguments{
\item{deseq_dataset}{}

\item{filt_method}{}

\item{min_counts}{}

\item{verbose}{}

\item{show_mltdnsty}{}

\item{cyclic_loess}{}
}
\value{
a DESeq2 object and a plot
}
\description{
Filter lowly abundant genes with the possibility of doing a normalisation on the genes to correct for trended biases.
This function will not estimate size factors of a DESeq2 object, but it will add normalization factors
to the DESeq2 object if the cyclic_loess is TRUE.
Plotting options allow to see how the read distribution changes using different filtering and normalisation parameters
}
\examples{
dds_fltrd <- filter_dds(dds, filt_method = "mean")
}
