% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils_plot.R
\name{find_containing_bin}
\alias{find_containing_bin}
\title{Return the bin location of a num given a range from start until end and a set number of bins.}
\usage{
find_containing_bin(start, end, num, num_bins = 50, return_full_df = F)
}
\arguments{
\item{start}{}

\item{end}{}

\item{num}{}

\item{num_bins}{}

\item{return_full_df}{}
}
\value{
A number or a dataframe
}
\description{
A short description...
}
\details{
If the number (\code{num}) is not between the start and end (\code{num > start & num <= end}), the function returns \code{NA}.
}
\examples{
# Example 1
find_containing_bin(start = 5515473, end = 5515517, pos = 5515489, return_full_df = F)
# Example 2: bin changes when pos increase
a <- 10 ; b <- 95
sapply(c(1:30), function(x) { find_containing_bin(start = a, end = b, pos = a+x) })
# Example 3: returning NA when num not in the range
find_containing_bin(start = 50, end = 150, num = 40, return_full_df = F)
}
