% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{gene_name_2_ensembl_id}
\alias{gene_name_2_ensembl_id}
\title{Convenient converter from gene name to ensembl ID}
\usage{
gene_name_2_ensembl_id(
  gene_name,
  only_ensembl_id = TRUE,
  verbose = FALSE,
  mRt_objct = ensembl
)
}
\arguments{
\item{gene_name}{A character specifying the gene name}

\item{only_ensembl_id}{Logical, whether to return only the gene name or a \code{data.frame} with more info. Default \code{TRUE}}

\item{verbose}{Logical. Do you want me to be chatty?}

\item{mRt_objct}{A \code{biomaRt} object created with the \code{gimme_mart()} function.}
}
\value{
A character corresponding to an ensembl gene ID or a \code{data.frame}.
}
\description{
The ENSEMBL gene IDs are returned in the same order as the input gene names.
}
\examples{
ensembl <- gimme_mart()
gene_name_2_ensembl_id("TSPAN6")
"ENSG00000000003"

# Example with mouse ensembl
ensembl <- gimme_mart(species = "mmusculus", latest = T, use_mirror = F, 
                      out_dir = F, verbose = T)
gene_name_2_ensembl_id(gene_name = "TSPAN6")
"ENSMUSG00000067377"
}
