% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{get_alphabet_palette}
\alias{get_alphabet_palette}
\title{Given an alphabet name set a palette for the plots.}
\usage{
get_alphabet_palette(alphabet, alphabet_name)
}
\arguments{
\item{alphabet}{A character vector with the alphabet letters.}

\item{alphabet_name}{The alphabet name returned by \code{guess_alphabet()}.}
}
\value{
An palette for \code{ggseqlogo}
}
\description{
Given an alphabet name set a palette for the plots.
}
