% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprofile_utils.R
\name{get_hPTM_FC}
\alias{get_hPTM_FC}
\title{Calculate summary stats across all samples for each hPTM and calculate hPTM are log2 Fold Change relative to control sample}
\usage{
get_hPTM_FC(data, Pseudocount = 1, Reference_Sample_Name)
}
\arguments{
\item{data}{Tibble imported with \code{read_EpiProfile_histone_ratio()}.}

\item{Pseudocount}{Number. By default adds 1 pseudo count to the \code{Area} values to avoid diving by zero.}

\item{Reference_Sample_Name}{name of the reference sample to be used for area log2 FC (e.g. WT/untreated condition)}
}
\value{
A tibble
}
\description{
Calculate summary stats across all samples for each hPTM and calculate hPTM are log2 Fold Change relative to control sample
}
\examples{
read_EpiProfile_histone_ratio("path/to/histone_ratio.xls") |>
    tidy_hPTMs() |>
    get_hPTM_FC(Reference_Sample_Name = 'WT_1_DIA')
}
