% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{get_mouse_tissue_devel_tbl}
\alias{get_mouse_tissue_devel_tbl}
\title{Wrapper function to import pre-processed vast-tool analysis tables of Mouse Development ENCODE data.}
\usage{
get_mouse_tissue_devel_tbl(
  inclusion_tbl = NULL,
  vst_expression_tbl = NULL,
  metadata_path = NULL,
  ensembl_gene_id,
  vst_id,
  filter_tbl = TRUE
)
}
\arguments{
\item{inclusion_tbl}{Path to vast-tools inclusion table that contains a \code{vst_id} event. Use default (\code{NULL}) to use pre-defined PSI table.}

\item{vst_expression_tbl}{Path to a vast-tools expression table (either \code{cRPKM} or \code{TPM}) that contains a \code{ensembl_gene_id} gene. Use default (\code{NULL}) to use pre-defined normalised TPMs table.}

\item{metadata_path}{A path to a metadata table. Use default (\code{NULL}) to use pre-defined metadata table.}

\item{ensembl_gene_id}{A valid Mouse ENSEMBL gene ID.}

\item{vst_id}{A valid Mouse Vast-tools ID.}

\item{filter_tbl}{Whether you want the full table to the filtered one with all the data.}
}
\value{
A tibble
}
\description{
Wrapper function to import pre-processed vast-tool analysis tables of Mouse Development ENCODE data.
}
\examples{
# Get Pax6 gene expression and Pax6 exon 6 PSI across mouse tissues development.
get_mouse_tissue_devel_tbl(ensembl_gene_id = "ENSMUSG00000027168", vst_id = "MmuEX0033804") 
 
}
