% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{gimme_mart}
\alias{gimme_mart}
\title{Handy function to create a biomaRt R object based on the ENSEMBL database to be used when querying biomart}
\usage{
gimme_mart(
  species = "hsapiens",
  version = "latest",
  verbose = TRUE,
  out_dir = FALSE,
  use_mirror = FALSE,
  which_mirror
)
}
\arguments{
\item{species}{a character specifying the species to work with. Currently the supported species are:
\itemize{
\item{\code{hsapiens} - Human (Homo Sapiens) Default}
\item{\code{mmulatta} - Rhesus Macaque (Macaca mulatta)}
\item{\code{mmusculus} - Mouse (Mus Musculus)}
\item{\code{rnorvegicus} - Rat (Rattus Norvegicus)}
\item{\code{btaurus} - Cattle, aka Cows & Bulls (Bos Taurus)}
\item{\code{mdomestica} - Gray short-tailed Opossum (Monodelphis domestica)}
\item{\code{ggallus} - Chicken (Gallus Gallus)}
}}

\item{version}{Which database version of biomart do you want? Use:
\itemize{
\item{\code{latest} - for the latest version of the database. }
\item{\code{penultimate} - for the one before the latest version of the database. }
\item{\code{number} - a specific version of the biomart database. If the version you query is not available an error is returned.}
}}

\item{verbose}{Logical. Print the version and genome assembly for the species queried. Default \code{TRUE}.}

\item{out_dir}{Either \code{FALSE} (default) or a character specifying a path where to save the biomaRt object. If directory does not exist it will be created.}

\item{use_mirror}{Logical. Do you want to use a mirror for creating the biomart object? Default \code{FALSE}. If \code{TRUE} the parameter version will be converted to \code{TRUE}.}

\item{which_mirror}{Specify which biomart mirror to use. Chose from:
\itemize{
\item{\code{www} - UK (Sanger Institute). in Cambridge, default}
\item{\code{uswest} - US West (Amazon AWS) on West Coast of US of America}
\item{\code{useast} - US East (Amazon AWS) on East Coast of US of America}
\item{\code{asia} - Asia (Amazon AWS) in Singapore}
}}
}
\value{
A biomaRt ensembl object
}
\description{
A function to get a biomaRt object during an R analysis that can later on be used for retrieving gene IDs, coordinates and more.
}
\note{
The latest ensembl release version of the mouse genome GRCm38 (aka mm10) is \code{102}.
}
\examples{
# Human ENSEMBL biomaRt object
human_ensembl <- gimme_mart()
}
