% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{gimme_psi_mat}
\alias{gimme_psi_mat}
\title{Extract the PSI from an inclusion table and return it as a matrix}
\usage{
gimme_psi_mat(inclusion_tbl, vst_id, quality_thrshld = "N", verbose = FALSE)
}
\arguments{
\item{inclusion_tbl}{path to vast-tools inclusion table that contains a vst_id event.}

\item{vst_id}{vast-tools alternative splicing event to grep in the \code{inclusion_tbl}.}

\item{quality_thrshld}{vast-tools event quantification quality score threshold. Must be one of "N", "VLOW", "LOW", "OK", "SOK". For more info read the official documentation \href{https://github.com/vastgroup/vast-tools#combine-output-format}{here} under "Column 8, score 1".}

\item{verbose}{Print out extra info.}
}
\value{
A matrix
}
\description{
The matrix contains the samples as row names and the vastID as column name
}
\examples{
gimme_psi_mat(inclusion_tbl = /path/to/inclusion/table/file.tab,
              vst_id = "HsaEX0000001")
}
