% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{grep_gene_expression}
\alias{grep_gene_expression}
\title{Grep the expression values of a gene from a vast-tools output table.}
\usage{
grep_gene_expression(
  vst_expression_tbl,
  ensembl_gene_id,
  tmp_dir = tempdir(),
  verbose = FALSE,
  clear_tmp = TRUE,
  fast_grep = TRUE
)
}
\arguments{
\item{vst_expression_tbl}{Path to a vast-tools expression table (either cRPKM or TPM)}

\item{ensembl_gene_id}{an ensembl gene id that is used to grep a line}

\item{tmp_dir}{a temporary directory where to same the intermediate files}

\item{verbose}{Logical, print some info?}

\item{clear_tmp}{Logical, remove the temporary files after the function is done?}

\item{fast_grep}{Logical, if TRUE use 'grep -m1' to stop the search after first match found. If FALSE all vst_expression_tbl is read.}
}
\value{
a data.frame
}
\description{
Grep the expression values of a gene from a vast-tools output table.
}
\examples{
grep_gene_expression(path_to_vst_Expression_tbl, ensembl_gene_id) |>
    tidy_vst_expr(ID_cols = "ensembl_geneid", expression_unit = "TMP")
}
