% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{grep_psi}
\alias{grep_psi}
\title{Grep the PSI values of an as event from a vast-tools output table.}
\usage{
grep_psi(
  inclusion_tbl,
  vst_id,
  tmp_dir = tempdir(),
  verbose = FALSE,
  clear_tmp = TRUE,
  fast_grep = TRUE,
  split_temp = TRUE
)
}
\arguments{
\item{inclusion_tbl}{path to vast-tools inclusion table that contains a vst_id event}

\item{vst_id}{vast-tools alternative splicing event to grep in the inclusion_tbl}

\item{tmp_dir}{path to a temporary location folder}

\item{clear_tmp}{Logical, remove the temporary files after the function is done?}

\item{fast_grep}{Logical, if TRUE use 'grep -m1' to stop the search after first match found. If FALSE all vst_expression_tbl is read.}

\item{split_temp}{Logical, if TRUE first create a temp header file then a temp AS event file and merge them. If FALSE create one temp file with both the header and the AS event PSI and quality scores.}
}
\value{
a data.frame
}
\description{
Grep the PSI values of an as event from a vast-tools output table.
}
\examples{
#' grep_psi(path_to_vst_tbl, vst_id) |>
    tidy_vst_psi()
}
