% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_files_utils.R
\name{gtf_attributes_as_list}
\alias{gtf_attributes_as_list}
\title{Helper function to handle the last column of a GTF file.}
\usage{
gtf_attributes_as_list(data)
}
\arguments{
\item{data}{A data frame of a GTF file with the 9th column labelled 'attribute'}
}
\value{
A tibble
}
\description{
Helper function to handle the last column of a GTF file.
}
\details{
This function helps convert the "attributes" column of a GTF file from character to a named list.
}
\examples{
gtf <- read_delim(file = path, delim = '\t', comment = '#!',
                  col_names = c('seqname', 'source', 'feature', 'start', 
                  'end', 'score', 'strand', 'frame', 'attribute'), 
                  col_types = c('i', 'c', 'c', 'i', 'i', 'c', 'c', 'c', 'c'),
                  n_max = 100) |>
                  gtf_attributes_as_list()
}
