% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{guess_alphabet}
\alias{guess_alphabet}
\title{Guess an alphabet type in a dataframe sequences.}
\usage{
guess_alphabet(
  data,
  ID_col,
  include_gaps = TRUE,
  name = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame with a minimum of 2 columns. One named \code{Sequence}, the other named as you prefer that will be specified with \code{ID_col}.}

\item{ID_col}{The name of the column in \code{data} to be used as the identifier of the \code{Sequence} column.}

\item{include_gaps}{Logical, include also the gap character '-' (dash) in the guess alphabet regex. Default \code{TRUE}.}

\item{name}{Logical, return only the guessed alphabet name instead of the letters in that alphabet. Default \code{FALSE}.}

\item{verbose}{Logical, print some info messages. Default \code{FALSE}.}
}
\value{
A character vector with the letters of that alphabet
}
\description{
It can can recognise sequences that are DNA/RNA; upper, lower, or both UPPER&lower case; gapped or non-gapped.
}
