% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{guess_species}
\alias{guess_species}
\title{Given a vast-tools AS event ID, return the species name.}
\usage{
guess_species(vastid, latin_name = F)
}
\arguments{
\item{vastid}{An alternative splicing event name from vast-tools.}

\item{latin_name}{Logical, whether or not you want the scientific name of the species.}
}
\value{
A character
}
\description{
Given a vast-tools AS event ID, return the species name.
}
\examples{
guess_species(vastid = "HsaEX000001")
# Human
}
