% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils_plot.R
\name{hallgrimskirkja_plot}
\alias{hallgrimskirkja_plot}
\title{Title}
\usage{
hallgrimskirkja_plot(
  df,
  left_slop,
  right_slop,
  left_bins_num = 50,
  centre_bins_num = 50,
  right_bins_num = 25,
  facet_regions = F,
  legend_position = c(0.94, 0.7),
  fill_by = c("region", "event", "bin", "count"),
  x_labels = c("exon", "generic", "bins"),
  event_type = NULL
)
}
\arguments{
\item{df}{}

\item{left_slop}{A number defining how much you slopped the bed file upstream (strand aware).}

\item{right_slop}{A number defining how much you slopped the bed file downstream (strand aware).}

\item{left_bins_num}{Number of bins you used for binning the upstream region.}

\item{centre_bins_num}{Number of bins you used for binning the central region(s).}

\item{right_bins_num}{Number of bins you used for binning the downstream region.}

\item{facet_regions}{Logical defining whether to slightly separate (facet) the individual regions. Can be either \code{TRUE} or \code{FALSE}.}

\item{legend_position}{Where to put the legend. Goes to \code{legend.postion} of the \code{ggplot2::theme()}.}

\item{fill_by}{How to colour the histogram, one of \code{region}, \code{event}, \code{bin}, \code{count}. See details for explanation.}

\item{x_labels}{How to label the X-axis text, still not perfect.}

\item{event_type}{Text to put in the title.}
}
\value{
A ggplot
}
\description{
Title
}
\details{
The parameter \code{fill_by} defines how to colour the plot: \code{region} colours by the type of region, meaning upstream, centre or downstream region.
\code{event} colours by the number of events (i.e. exons) with an SNP in that specific bin. It serves to check that high bars are due to many different event regions and not one single hot spot in few events.
\code{bin} colours by the bin number in each region and \code{count} fills based on how tall is the bar in the histogram.
}
\note{
save the plots with ggsave( device = cairo_pdf) and not with device = 'pdf'.
}
