% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_plot.R
\name{hist_dPSI}
\alias{hist_dPSI}
\title{Plot the dPSI as an histogram using ggplot2}
\usage{
hist_dPSI(data, by_complex = FALSE)
}
\arguments{
\item{data}{A data.frame that has at least one column called 'dPSI'}

\item{by_complex}{Logical, whether to split the histogram by AS event complexity/type. Uses vast-tools tables "COMPLEX" column.}
}
\value{
A ggplot2 plot
}
\description{
Plot the dPSI as an histogram using ggplot2
}
\examples{
read_vst_tbl(path = 'path/to/compare/inclusion/tbl', show_col_types = FALSE) \%>\%
   tidy_vst_psi(verbose = FALSE) \%>\%
   hist_dPSI(by_complex = TRUE)
}
