% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{inclusion_tbl2bed}
\alias{inclusion_tbl2bed}
\title{Create a bed file from events in a vast-tools inclusion table}
\usage{
inclusion_tbl2bed(
  path,
  header = TRUE,
  remove_chr = FALSE,
  out_bed_name,
  out_path,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to vast-tools inclusion file or a tab delimited file with the same first ID columns as vast-tools inclusion tables.}

\item{header}{Does the vast-tools inclusion file have a header? If \code{FALSE} consider first row as an event.}

\item{remove_chr}{If \code{TRUE} remove 'chr' from the first column of the bed files.}

\item{out_bed_name}{Name of output bed file, without the \code{.bed} extension. If missing use input file basename.}

\item{out_path}{Path to directory where to save bed file. If folder doesn't exist create it.}

\item{verbose}{If \code{TRUE} print now many lines are in the bed file.}
}
\value{
a sorted bed file without a header with EVENT ID in the 4th column and gene name in the 5th column.
}
\description{
The input file (specified with path), doesn't need to be specifically a vast-tools output file.
As long as the first five columns of the input file match the structure of a vast-tools inclusion table, this function will work.
}
\details{
This function can also report the strand for exons, in the future I might implement this also for introns.
}
\examples{
inclusion_tbl2bed(path = vst_tbl_path, header = T, remove_chr = T, out_path = bed_dir)
}
