% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{longest_common_prefix}
\alias{longest_common_prefix}
\title{Find the longest common prefix between 2 random names. If no common prefix is found the same input values are returned.}
\usage{
longest_common_prefix(names, uniquify = TRUE, verbose = TRUE)
}
\arguments{
\item{names}{A character vector with length higher than 1.}

\item{uniquify}{Logical, should \code{names} become a unique vector before sampling 2 random \code{names}? Default \code{FALSE}, setting to \code{TRUE}, could help finding a better longest common prefix.}

\item{verbose}{Logical, print common prefix found.}
}
\value{
A character vector without common prefix that is the same length as the input.
}
\description{
Find the longest common prefix between 2 random names. If no common prefix is found the same input values are returned.
}
\examples{
new_names <- longest_common_prefix(names)
}
