% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_plot.R
\name{plot_best_corr_genes}
\alias{plot_best_corr_genes}
\title{Plot to check the top and bottom correlating genes of one AS event PSI vs all genes expression levels. (1 vs many)}
\usage{
plot_best_corr_genes(data, ...)
}
\arguments{
\item{data}{A tibble generated with \verb{all_gene_expr_corr(num_genes = <NUM>, map_ID_2_names = T)}}

\item{...}{Extra parameters passed to \code{geom_point}}
}
\value{
A ggplot2 plot.
}
\description{
Plot to check the top and bottom correlating genes of one AS event PSI vs all genes expression levels. (1 vs many)
}
\examples{
#' all_gene_expr_corr(inclusion_tbl = psi_path, vst_id = "HsaEX0000001", 
                   vst_expression_tbl = expr_path, corr_method = "spearman", 
                   verbose = F, use = "complete.obs",
                   quality_thrshld = "LOW",
                   num_genes = 10, expr_min_mean_fltr = 10,
                   map_ID_2_names = T, species = "hsapiens") |>
                   plot_best_corr_genes(size = 5, stroke = 0.25)  
}
