% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_plot.R
\name{plot_corr_gene_expr_psi}
\alias{plot_corr_gene_expr_psi}
\title{Plot gene expression vs alternatively spliced event PSI}
\usage{
plot_corr_gene_expr_psi(
  data,
  quality_thrshld = "N",
  external_gene_name,
  vastid,
  unit,
  text = FALSE,
  beautify = FALSE,
  xzero = TRUE,
  colour = c("score", "PSI"),
  save_plot = FALSE,
  out_plot_dir,
  verbose = TRUE,
  return_data = FALSE,
  subttl = NULL
)
}
\arguments{
\item{data}{}

\item{quality_thrshld}{}

\item{external_gene_name}{}

\item{vastid}{}

\item{unit}{}

\item{text}{}

\item{beautify}{}

\item{xzero}{}

\item{colour}{}

\item{save_plot}{}

\item{out_plot_dir}{}

\item{verbose}{}

\item{return_data}{}

\item{subttl}{}
}
\value{

}
\description{
Plot gene expression vs alternatively spliced event PSI
}
