% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{plot_JSD_logo}
\alias{plot_JSD_logo}
\title{Calculate and plot the Jensen-Shannon divergence between 2 sets of sequences contained in 2 data.frames}
\usage{
plot_JSD_logo(
  df1,
  df2,
  ID_col,
  alphabet,
  y_lims = c(-0.5, 0.5),
  max_JS_div_thrshld,
  anno_width = 0.75,
  highlight_colour = "grey74",
  axis_txt_size = 10,
  ttl_txt = NULL,
  ...
)
}
\arguments{
\item{df1}{A data.frame with a column called \code{Sequence} and anther defined in \code{ID_col}. Use the wrapper \code{fasta2df()} to import a fasta file to a dataframe that is compatible with this function.}

\item{df2}{A different data.frame with the same properties as \code{df1}.}

\item{ID_col}{A character specifying the column name in \code{df1} and \code{df2} to be used as the sequence ID.}

\item{alphabet}{A character vector containing the alphabet letters present in \code{Sequence}. Guessed by default from \code{df1}.}

\item{y_lims}{A numeric vector of length 2 specifying the Y-axis min and max value. Default \code{c(-0.5, 0.5)}}

\item{max_JS_div_thrshld}{The maximal absolute JS divergence threshold to consider a letter position to be highlighted by a vertical bar in the plot.}

\item{anno_width}{A small number that defines how wide the vertical bar should be. Default \code{0.75}.}

\item{highlight_colour}{A colour name to fill the letter highlighting rectangle. Default \code{grey74} with \code{alpha = 0.5}.}

\item{axis_txt_size}{A number specifying the size of the axis text in the plot. Default 10.}

\item{ttl_txt}{Some text in quotes specifying the plot title.}

\item{...}{Advanced parameters passed to \link[niar]{renumber_logo_seq_breaks}.}
}
\value{
A ggplot sequence logo with the Jensen-Shannon divergence on the Y-axis.
}
\description{
Calculate and plot the Jensen-Shannon divergence between 2 sets of sequences contained in 2 data.frames
}
\examples{
plot_JSD_logo(df1 = clade1, df2 = clade2f, ID_col = 'Species',
              max_JS_div_thrshld = 0.2, anno_width = 0.5)
}
