% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{plot_pwm_as_hm}
\alias{plot_pwm_as_hm}
\title{Visualise a positional weight matrix as a heatmap}
\usage{
plot_pwm_as_hm(long_df, cell_txt_size = 3, axis_txt_size = 10)
}
\arguments{
\item{long_df}{A data.frame obtained with \code{df2pwd(long_format = TRUE)}}

\item{cell_txt_size}{Text size of the heatmap cells that corresponds to the letter weight in that position. Default 3.}

\item{axis_txt_size}{Text size of axis labels. Default 10.}
}
\value{
A ggplot plot
}
\description{
Visualise a positional weight matrix as a heatmap
}
\examples{
df2pwm(data, ID_col = 'Species', alphabet = c('a', 'c', 'g', 't'), long_format = T) |>
  plot_pwm_as_hm() 
}
