% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rmats.R
\name{plot_rmats_summary}
\alias{plot_rmats_summary}
\title{Plot the number of AS events quantified with rMATS-turbo}
\usage{
plot_rmats_summary(data, legend_position = c(0.8, 0.8))
}
\arguments{
\item{data}{A data.frame generated with \code{read_rmats_summary()}}

\item{legend_position}{The relative location of the legend in the plot. Passed to \code{theme(legend.position.inside)}}
}
\value{
A ggplot
}
\description{
Plot the number of AS events quantified with rMATS-turbo
}
\examples{
# read_rmats_summary(outdir = "path/to/rmats/output" ) |>
#          plot_rmats_summary()
}
\seealso{
\code{\link[=read_rmats_summary]{read_rmats_summary()}}
}
