% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{plot_rudimentary_JSD_logo}
\alias{plot_rudimentary_JSD_logo}
\title{Quick visualisation of the Jensen-Shannon divergence}
\usage{
plot_rudimentary_JSD_logo(
  long_df,
  axis_txt_size = 10,
  alphabet_palette = c(t = "forestgreen", c = "dodgerblue", g = "orange", a =
    "firebrick3")
)
}
\arguments{
\item{long_df}{A data.frame created with \code{df2JSD(long_format = T)}.}

\item{axis_txt_size}{}

\item{alphabet_palette}{}
}
\value{
A ggplot plot
}
\description{
Quick visualisation of the Jensen-Shannon divergence
}
\examples{
df2JSD(df1 = data_p, df2 = data_q, ID_col, alphabet, long_format = T) |>
  plot_rudimentary_JSdiv_logo()
# alphabet is case sensitive, in this example the lowercase letters will be in grey.
df2JSD(df1 = df_clade1, df2 = df_clade2, ID_col = 'Species', 
                alphabet = c('A', 'C', 'T', 'G', 'a', 'c', 't', 'g'), long_format = T) |>
   plot_rudimentary_JSD_logo(alphabet_palette = c('A' ="firebrick3", 'C' = "dodgerblue", 'T' = "forestgreen", 'G' ="orange") )
}
