% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{plot_rudimentary_logo}
\alias{plot_rudimentary_logo}
\title{Quick visualisation of a logo. Not a fancy plot, just useful for quick checks.}
\usage{
plot_rudimentary_logo(
  long_df,
  axis_txt_size = 10,
  alphabet_palette = c(t = "forestgreen", c = "dodgerblue", g = "orange", a =
    "firebrick3")
)
}
\arguments{
\item{long_df}{A data.frame obtained with \code{df2bits(long_format = TRUE)}.}

\item{axis_txt_size}{Text size of axis labels. Default 10.}

\item{cell_txt_size}{Text size of the heatmap cells that corresponds to the letter weight in that position. Default 3.}
}
\value{
A ggplot plot
}
\description{
If you're looking a good way to make logos check out the package \href{https://omarwagih.github.io/ggseqlogo/}{ggseqlogo}
}
\examples{
df2bits(data, ID_col = 'Species', alphabet = c('a', 'c', 'g', 't'), small_n_correction = FALSE, long_format = T) |>
plot_rudimentary_logo()

df2bits(data, ID_col = 'Species', alphabet = c('a', 'c', 'g', 't'), small_n_correction = TRUE, long_format = T ) |>
plot_pwm_as_bars()
}
