% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_plot.R
\name{plot_samples_PSI}
\alias{plot_samples_PSI}
\title{Plot the PSI of all samples in a data.frame}
\usage{
plot_samples_PSI(data, simplify_names = TRUE)
}
\arguments{
\item{data}{A dataframe with certain columns. Do NOT use this function for plotting a lot of AS events.}

\item{simplify_names}{Logical, Should names be shortened? Helps with shorter X-axis text names.}
}
\value{
A ggplot2 plot
}
\description{
Plot the PSI of all samples in a data.frame
}
\examples{
read_vst_tbl(path = "file/inclusion/table.tab", show_col_types = FALSE) \%>\%
     tidy_vst_psi(verbose = F) \%>\%
     subset(abs(dPSI) >= 80 ) \%>\%
     plot_samples_PSI()
}
