% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprofile_plot.R
\name{plotALLU_hPTM}
\alias{plotALLU_hPTM}
\title{Make an alluvial plot with the histone PTMs ratio in each sample}
\usage{
plotALLU_hPTM(data, peptide_name, right_sample_name, num_col = 12)
}
\arguments{
\item{data}{}

\item{peptide_name}{Name of the peptide to plot}

\item{right_sample_name}{name of the sample on the RHS that will have some extra geom_labels}

\item{num_col}{number of colours to use for the plot}
}
\value{
A nice alluvial plot
}
\description{
Make an alluvial plot with the histone PTMs ratio in each sample
}
\examples{
    tidy_hPTMs() |>
    get_hPTM_FC(Reference_Sample_Name = 'WT_1_DIA') |>
    plotALLU_hPTM(peptide_name = "KSAPATGGVKKPHR(H3_27_40)",
                  right_sample_name = "Last_Sample_DIA", 
                  num_col = 12)
}
