% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprofile_plot.R
\name{plotMA_hPTM}
\alias{plotMA_hPTM}
\title{Plot the histone PTMs fold changes as an MA plot.}
\usage{
plotMA_hPTM(
  data,
  sample_name,
  RT_min_sd_thrshld = 2,
  fc_thrshld = 2,
  area_max_thrshld = 2e+12,
  label_txt_size = 2,
  Pseudocount_value = 1
)
}
\arguments{
\item{data}{A tibble processed with \code{get_hPTM_FC()}.}

\item{sample_name}{Character. Name of the sample to plot.}

\item{RT_min_sd_thrshld}{Numeric. A minimum threshold for the retention time standard deviation across all samples in the tibble for an individual PTM. Defatult \code{2} min.}

\item{fc_thrshld}{Numeric. Label histone PTMs in the plot with absolute log2 fold change higher than \code{fc_thrshld}. Default \code{2}.}

\item{area_max_thrshld}{Numeric. Label histone PTMs in the plot with with average area across samples higher higher than \code{area_max_thrshld}. Default \code{2e12}.}

\item{label_txt_size}{Numeric. Size of the labels in the plot.}

\item{Pseudocount_value}{Numeric. Pseudo count value used in \code{get_hPTM_FC()}, used only for plotting info, not actually added. Default \code{1}.}
}
\value{
A nice ggplot.
}
\description{
Plot the histone PTMs fold changes as an MA plot.
}
\details{
This plotting function is inspired by bulk mRNA-seq MA plots as in DESeq2 analysis. Histone PTMs are coloured following Karolin Luger's histone proteins palette.
}
\examples{
read_EpiProfile_histone_ratio("path/to/histone_ratio.xls") |>
    tidy_hPTMs() |>
    get_hPTM_FC(Reference_Sample_Name = 'WT_1_DIA') |>
    plotMA_hPTM(data = subset(RT_sd <= 2 & Ratio > 0.000004 & Area != 1),
                sample_name = "KO_2_DIA")
}
