% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{print_seq_perm}
\alias{print_seq_perm}
\title{Pretty print a sequences permutation data frame}
\usage{
print_seq_perm(perm_dat, alphabet = c("A", "C", "G", "T"), k = 6)
}
\arguments{
\item{perm_dat}{the data frame generated with \code{permutate_seq()}}

\item{alphabet}{Letters to permutate in order to build different sequences}

\item{k}{Pretty print option: show k rows per letter of the alphabet}
}
\value{
A data frame slicing the permutation data by letter
}
\description{
Show the first k rows of all the words starting with each letter of the \code{alphabet}.
}
\details{
In addition to the first k words of each letter also the last k rows are always returned.
}
\seealso{
\link{permutate_seq}
}
