% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_files_utils.R
\name{read_bed}
\alias{read_bed}
\title{Import a bed file into R as a tibble}
\usage{
read_bed(path, header)
}
\arguments{
\item{path}{Path to bed file in your system specified as character}

\item{header}{Is the bed file first line specifying the column names?}
}
\value{
A tibble
}
\description{
Import a bed file into R as a tibble
}
\details{
If header = \code{F} than the UCSC official bed files column names are added.
}
\examples{
read_bed(path = 'path/to/bed/file', header = F)
}
