% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_files_utils.R
\name{read_gtf}
\alias{read_gtf}
\title{Import a GTF as tibble}
\usage{
read_gtf(path, max_rows = 10000, col_attribute_type = "character", ...)
}
\arguments{
\item{max_rows}{The maximum rows to import. Helps avoiding memory issues. Default 10000.}

\item{col_attribute_type}{Specify how to handle the last column of a GTF file. See details.
#' \itemize{
\item{\code{character} - Default. Basically do nothing. }
\item{\code{list} - Convert the column to a list }
\item{\code{split} - Split the column into different separate columns}
}}

\item{...}{Additional parameters to pass to \code{read_delim()}}

\item{A}{valid path to a GTF file.}
}
\value{
A GTF file as a tibble
}
\description{
Import a genomics GTF file and offer different way to handle the last 9th column (with a bunch of extra info).
}
\details{
The parameter \code{col_attribute_type} can be either: "character", "list", or "split".
The first option (\code{character}) imports the data as is.
The option \code{list} coerce the column into a list column in the data frame.
The last option \code{split} divides all the attributes into separate columns with a header.
}
\examples{
hg38gtf <- read_gtf(path = hg38_gtf_path)
}
