% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rmats.R
\name{read_rmats}
\alias{read_rmats}
\title{Read rMATS Output Files and Process Splicing Event Data}
\usage{
read_rmats(
  outdir,
  reads_counts_method = c("JC", "JCEC"),
  return = c("SHARED_COLS_ONLY", "SE", "ALT53SS", "MXE", "RI", "ALL"),
  max_lines = 1e+09
)
}
\arguments{
\item{outdir}{Character. Path/to/the/directory/containing/rMATS/output/files.}

\item{reads_counts_method}{Character. The method to use for reading counts. Either \code{JC} (default) for junction counts or \code{JCEC} for junction counts plus exon counts.}

\item{return}{Character. Specifies the type of data to return. Options are \code{SHARED_COLS_ONLY} (default), \code{SE}, \code{ALT53SS}, \code{MXE}, \code{RI}, and \code{ALL}.}

\item{max_lines}{Numeric. Maximum number of lines to read from each file. Default is \code{1e9}.}
}
\value{
A data frame containing the processed splicing event data. The structure of the returned data frame depends on the value of the \code{return} parameter. If \code{return} is set to \code{ALL} a list of dataframes is returned.
}
\description{
This function reads and processes output files generated by the rMATS (Replicate Multivariate Analysis of Transcript Splicing) software.
It imports the data for different types of splicing events, scales the PSI values, and returns a combined data frame with the specified columns (if needed).
}
\details{
The function processes five types of splicing events: Skipped Exons (SE), Alternative 5' Splice Sites (A5SS), Alternative 3' Splice Sites (A3SS), Mutually Exclusive Exons (MXE), and Retained Introns (RI).
It reads the corresponding files, scales the PSI values, and returns a tibble.
The user can choose to return only the shared columns across all event types or data specific to a particular event type with \code{return}.
To understand the different between \code{JC} and \code{JCEC} please refer to the rMATS manual.
}
\examples{
# result <- read_rmats(outdir = "path/to/rmats/output", reads_counts_method = "JC", return = "SHARED_COLS_ONLY")
}
