% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rmats.R
\name{read_rmats_summary}
\alias{read_rmats_summary}
\title{Import the rMATS-turbo summary files.}
\usage{
read_rmats_summary(outdir)
}
\arguments{
\item{outdir}{Character. Path/to/the/directory/containing/rMATS/output/files.}
}
\value{
A tibble
}
\description{
This function imports the numbers of total and significant AS events quantified with both JC and JCEC.
}
\examples{
# read_rmats_summary(outdir = "path/to/rmats/output" ) |>
#          plot_rmats_summary()
}
\seealso{
\code{\link[=plot_rmats_summary]{plot_rmats_summary()}}
}
