% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{read_vst_tbl}
\alias{read_vst_tbl}
\title{Wrapper for readr::read_delim for a tab delimited table. Fastest way to read this kind of data into \code{R}.}
\usage{
read_vst_tbl(path, verbose = FALSE, ...)
}
\arguments{
\item{path}{A character vector providing the 'path/to/the/table/to/read'.}

\item{verbose}{Logical. Use \code{TRUE} to know the table file size being read into \code{R}.}

\item{...}{extra info to pass to \code{read_delim}.}
}
\value{
A tibble
}
\description{
Wrapper for readr::read_delim for a tab delimited table. Fastest way to read this kind of data into \code{R}.
}
\examples{
read_vst_tbl(path = "location/of/tab/delimited/file/inclusion/table.tab") |>
   tidy_vst_psi()
 
}
