% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprofile_utils.R
\name{Rename_EpiProfile_colnames}
\alias{Rename_EpiProfile_colnames}
\title{Auxiliary function to modify the EpiProfile columns names on the fly while importing data.}
\usage{
Rename_EpiProfile_colnames(col)
}
\arguments{
\item{col}{the columns to rename}
}
\description{
Auxiliary function to modify the EpiProfile columns names on the fly while importing data.
}
