% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{res2df}
\alias{res2df}
\title{Work in PROGRESS. Convert a general DEseq2 result object into a data frame.}
\usage{
res2df(
  res,
  map_gene_IDs = F,
  histone_PTMs = F,
  pval_cutoff = 0.05,
  log2FC_cutoff = 0
)
}
\arguments{
\item{res}{A DESeq2 object created with \code{results()} or \code{lfcShrink()}}

\item{map_gene_IDs}{Logical. Do you want to convert ensembl gene IDs to gene names?}

\item{histone_PTMs}{Logical. Are you analysing histone PTMs? Used to name the first column \code{First_Col} to be compatible with other functions.}

\item{pval_cutoff}{Numeric threshold to consider observations statistically significant. Default \code{0.05}.}

\item{log2FC_cutoff}{Numeric threshold to consider observations moving in a given direction (UP or DOWN). Default \code{0}.}
}
\value{
A tibble data frame
}
\description{
Work in PROGRESS. Convert a general DEseq2 result object into a data frame.
}
\examples{

# Somewhere in your DESeq2 analysis
resLFC <- lfcShrink(dds, coef="condition_treated_vs_untreated", type="apeglm")
res_as_df <- res2df(resLFC)
}
