% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{save_counts}
\alias{save_counts}
\title{Save the counts of a DESeq2 object.}
\usage{
save_counts(deseq_dataset, name, out_dir, tidy = TRUE, ...)
}
\arguments{
\item{deseq_dataset}{A dds}

\item{name}{A name for the output file}

\item{out_dir}{Path to the output directory}

\item{tidy}{Logical, whethere to reshape the data into a long (\code{TRUE}) or keep it as wide ('FALSE`) format.}

\item{...}{extra arguments passed to \code{dds2counts()} function, like \code{counts_are_genes = T|F} or \code{norm_counts = T|F}}
}
\value{
Nothing, this function writes a file
}
\description{
Save the counts of a DESeq2 object.
}
\examples{
save_counts(dds, name = "Your_comparison_long", out_dir = "/path/to/dir")
}
