% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{save_df_gsea_list}
\alias{save_df_gsea_list}
\title{This function is still experimental as it should be broken down into simpler sub-functions}
\usage{
save_df_gsea_list(
  res,
  name,
  map_df,
  baseMean_thrshold = 300,
  pval_adj_thrshold = 0.001,
  out_dir
)
}
\arguments{
\item{res}{}

\item{name}{}

\item{map_df}{}

\item{baseMean_thrshold}{}

\item{pval_adj_thrshold}{}

\item{out_dir}{}
}
\value{
Nothing, this function writes to file.
}
\description{
This function is still experimental as it should be broken down into simpler sub-functions
}
