% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_utils.R
\name{show_dds_counts_freq}
\alias{show_dds_counts_freq}
\title{Helper function to check distribution of reads in DESeq2 object}
\usage{
show_dds_counts_freq(
  deseq_dataset,
  xlim = c(0, 20),
  title = "Here goes the title",
  min_counts = 5
)
}
\arguments{
\item{deseq_dataset}{}

\item{xlim}{}

\item{title}{}

\item{min_counts}{Small number of pseudocounts added to the normalised counts before the \code{log2()}.}
}
\value{
A nice ggplot2 density plot
}
\description{
Helper function to check distribution of reads in DESeq2 object
}
\examples{
show_dds_counts_freq(dds)
}
