% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{showme_PCA2D}
\alias{showme_PCA2D}
\title{Given a matrix show me the principal components and other neat things.}
\usage{
showme_PCA2D(
  mat,
  x = 1,
  y = x + 1,
  mt,
  mcol,
  m_fill = mcol,
  m_label = mcol,
  n_top_var = 100,
  filt_mat = FALSE,
  NA_filt_thrshld = 0.95,
  show_variance = FALSE,
  show_stats = FALSE,
  n_loadings = NULL,
  return_data = FALSE,
  real_aspect_ratio = TRUE,
  ...
)
}
\arguments{
\item{mat}{A matrix}

\item{x}{Integer indicating principal component on X-axis. Defaul 1}

\item{y}{Integer indicating principal component on Y-axis. Default x + 1}

\item{mt}{Dataframe: contains matrix columns additional information, if not provided plot can't be labelled nor coloured according to specific information.}

\item{mcol}{Character pointing one column of metadata that contains the matrix column names}

\item{m_fill}{Character pointing one column of metadata specifying how to colour the samples.}

\item{m_label}{Character pointing one column of metadata specifying how to label the points in the PCA. Set to FALSE for omitting labels in plot.}

\item{n_top_var}{Integer: Number of most variable matrix rows to use for \code{prcomp}}

\item{filt_mat}{Logical. Fil}

\item{NA_filt_thrshld}{Integer between 0 and 1: maximum \% of NA accepted on \code{mat} rows.}

\item{show_variance}{Logical. Show an extra plot with the variance on each component.}

\item{show_stats}{Logical. Show an extra plot with the summary statistics for the data in \code{mat}.}

\item{n_loadings}{Integer indicating how many top and bottom loadings to plot}

\item{return_data}{Logical. If \code{TRUE} returns data for plot instead of plot.}

\item{...}{Set extra parameter for the \code{prcomp} function like \code{scale.} (default FALSE) and \code{center} (default TRUE).}

\item{real_aspect_raio}{Logical. If \code{TRUE} represent the distances between samples as faithfully as possible. Take into account that the second eigenvalue is always smaller than the first, sometimes considerably so, thus \code{TRUE} normalize the axis aspect ratio to the relevant eigenvalue ratio only of the PCA plot. Adapted from by: https://f1000research.com/articles/5-1492/v2}
}
\value{
Either a plot, a combination of plots or a dataframe.
}
\description{
Given a matrix show me the principal components and other neat things.
}
\examples{
showme_PCA2D(mat)

showme_PCA2D(mat = mat,  mt = mt, mcol = "sample_name", m_fill = "replicate",
             x = 3, show_stats = T, m_label = F)
             
showme_PCA2D(mat = mat, mt = mt, mcol = "sample_name", n_loadings = 12)              
}
