% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_files_utils.R
\name{slop_bed}
\alias{slop_bed}
\title{Extend upstream or downstream the coordinates of a bed file}
\usage{
slop_bed(bed, upstream = 10, downstream = 5, strand_aware = T)
}
\arguments{
\item{bed}{a tibble imported with \code{read_bed()}}

\item{upstream}{How many nucleotides to add upstream to the start coordinate}

\item{downstream}{How many nucleotides to add upstream to the end coordinate}

\item{strand_aware}{Logical, whether or not to slop in a strand aware fashion.}
}
\value{
A tibble
}
\description{
If \code{strand_aware} then for the negative strands start position is slopped by
the negative \code{downstream} value and
the end position is slopped by adding the upstream value.
}
\details{
strand can only be specified with \code{+} or \code{-},
}
\examples{
read_bed(path = test_coord, header = F)

# A tibble: 3 × 6
# chr   start   end name    score strand
# <chr> <dbl> <dbl> <chr>   <dbl> <chr> 
# chr10   100   200 test1     0 +     
# chr14   500   550 test2     0 -     
# chr1    150   190 test3     0 -   

read_bed(path = exons_coord, header = F) |> slop_bed(upstream = 10, downstream = 5, strand_aware = T)

# A tibble: 3 × 6
# chr   start   end name    score strand
# <chr> <dbl> <dbl> <chr>   <dbl> <chr> 
# chr10    90   205 test1     0 +     
# chr14   495   560 test2     0 -     
# chr1    145   200 test3     0 -     

}
