% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_utils.R
\name{test_regexes}
\alias{test_regexes}
\title{Handy wrapper to grep regexes on multiple vectors}
\usage{
test_regexes(chrctr, regex)
}
\arguments{
\item{chrctr}{A character vector}

\item{regex}{A character with PERL compatible regexes, case sensitive.}
}
\value{
A named logical vector
}
\description{
Handy wrapper to grep regexes on multiple vectors
}
