% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{tidy_vst_expr}
\alias{tidy_vst_expr}
\title{Reshape a wide vast-tools expression table to a long format}
\usage{
tidy_vst_expr(
  data,
  expression_unit = c("TPM", "cRPKM"),
  ID_cols = c("ID", "NAME", "Names", "ensembl_gene_id")
)
}
\arguments{
\item{data}{A dataframe generated with \code{grep_gene_expression} from a vast-tools gene expression table.}

\item{expression_unit}{Character: either \code{TPM} or \code{cRPKM} based on the gene expression unit in \code{data}.}

\item{ID_cols}{The columns names to use as IDs in \code{data}}
}
\value{
A reshaped \code{data.frame} in long format as \code{tibble}.
}
\description{
Reshape a wide vast-tools expression table to a long format
}
\examples{
grep_gene_expression(path_to_vst_Expression_tbl, ensembl_gene_id) |>
    tidy_vst_expr(ID_cols = "ensembl_geneid", expression_unit = "TMP")
}
