% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_utils.R
\name{tidy_vst_psi}
\alias{tidy_vst_psi}
\title{Reshape a wide vast-tools inclusion table into a tidyverse-friendly long format table}
\usage{
tidy_vst_psi(
  vst_psi_tbl,
  num_id_cols = 6,
  num_of_Score_Types = 5,
  quality_col_suffix = "-Q",
  return_quality_scores = TRUE,
  return_S1 = TRUE,
  verbose = TRUE,
  add_ID_col = FALSE,
  col_ID_name = "banana"
)
}
\arguments{
\item{vst_psi_tbl}{A dataframe generated with \code{grep_psi()} from a vast-tools inclusion table}

\item{num_id_cols}{How many first num_id_cols to consider as info/metadata/IDs of the AS event ID in the table}

\item{num_of_Score_Types}{How many quality scores are present in the columns with headers ending with quality_col_suffix}

\item{quality_col_suffix}{Suffix identifying the quality control columns. Default is "-Q".}

\item{return_quality_scores}{Logical. Do you want the individual scores to be returned in the output data.frame? Defaul true}

\item{return_S1}{Logical. Limit the score to be returned to only S1. Default T.}

\item{verbose}{Print out information}

\item{add_ID_col}{Logical. Do you want to add an extra col_ID_name to the output data.frame? Default false.}

\item{col_ID_name}{Extra column that could be used to add a new identifier to the data. Default is "banana".}
}
\value{
A data.frame in long format.
}
\description{
Reshape a wide vast-tools inclusion table into a tidyverse-friendly long format table
}
\examples{
grep_psi(path_to_vst_PSI_tbl, vst_id) \%>\%
    tidy_vst_psi()
}
