% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{gimme_mart}
\alias{gimme_mart}
\title{Handy function to create a biomaRt R object based on the ENSEMBL database to be used when querying biomart}
\usage{
gimme_mart(
  species = "hsapiens",
  latest = TRUE,
  use_mirror = FALSE,
  which_mirror,
  out_dir = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{species}{a character specifying the species to work with. Currently the supported species are:
\itemize{
\item{\code{hsapiens} - Human (Homo Sapiens) Default}
\item{\code{mmusculus} - Mouse (Mus Musculus)}
\item{\code{rnorvegicus} - Rat (Rattus Norvegicus)}
\item{\code{ggallus} - Chicken (Gallus Gallus)}
}}

\item{latest}{Logical. Do you want to use the latest biomart version? Default \code{TRUE}. If \code{FALSE} penultimate version is used.}

\item{use_mirror}{Logical. Do you want to use a mirror for creating the biomart object? Default \code{FALSE}.}

\item{which_mirror}{Specify which biomart mirror to use. Chose from:
\itemize{
\item{\code{www} - UK (Sanger Institute). in Cambridge, default}
\item{\code{uswest} - US West (Amazon AWS) on West Coast of US of America}
\item{\code{useast} - US East (Amazon AWS) on East Coast of US of America}
\item{\code{asia} - Asia (Amazon AWS) in Singapore}
}}

\item{out_dir}{Either \code{FALSE} (default) or a character specifying a path where to save the biomaRt object. If directory does not exist it will be created.}

\item{verbose}{Logical. Print info messages? Default \code{FALSE}.}
}
\description{
Handy function to create a biomaRt R object based on the ENSEMBL database to be used when querying biomart
}
\examples{
# Human ENSEMBL biomaRT object
ensembl <- gimme_mart()
}
