% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{longest_most_abundant_common_suffix}
\alias{longest_most_abundant_common_suffix}
\title{Find longest and most abundant common suffix}
\usage{
longest_most_abundant_common_suffix(
  x,
  k = 80,
  max_suffixes = 4,
  ignore.case = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A character vector}

\item{k}{An integer between 1 and 99. How much more abundant should the first suffix be relative to the second most abundant suffix?}

\item{max_suffixes}{An integer. If there are more than \code{max_suffixes} in \code{x} don't check for relative percentage abundance.}

\item{ignore.case}{Logical. Convert all \code{x} to upper case.}

\item{verbose}{Logical. Print out info.}
}
\value{

}
\description{
Find longest and most abundant common suffix
}
