% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_plot.R
\name{plot_mouse_tissue_devel}
\alias{plot_mouse_tissue_devel}
\title{Nice plot to check expression and PSI across ENCODE mouse developmental stages of different tissues.}
\usage{
plot_mouse_tissue_devel(
  data_tbl,
  title = NULL,
  legend = c("inside", "outside"),
  save_plot = FALSE,
  plot_name = NULL,
  out_plot_dir = NULL,
  width = 7,
  height = 7,
  colour_bar = c("BlueRed", "viridis"),
  PSI_limits = c(0, 100),
  blck_wht_PSI_col_thshld = 70
)
}
\arguments{
\item{data_tbl}{A tibble generated with \code{get_mouse_tissue_devel_tbl}.}

\item{title}{A title for the plot. By default (\code{NULL}) is "gene ~ vast-id".}

\item{legend}{Either \code{inside} or \code{outside} for specifying the position of the legend relative to the plot.}

\item{save_plot}{Logical, whether or not to save the plot as pdf.}

\item{plot_name}{Name of the pdf plot file. By default (\code{NULL}) is a meaningful name.}

\item{out_plot_dir}{Path specifying where to save the plot. By default (\code{NULL}) is dated subfolder.}

\item{width}{Width of the plot in cm (numeric). Defaults are already very good.}

\item{height}{Height of the plot in cm (numeric). Defaults are already very good.}

\item{colour_bar}{Colour palette to represent the PSI values. Select between \code{BlueRed} or \code{viridis}.}

\item{PSI_limits}{A vector of length 2, specifying the minimum and maximum PSI colour limits in the legend.}

\item{blck_wht_PSI_col_thshld}{The PSI text is shown in black, PSI values above this threshold will be displayed in white instead. Gives better contrast with dark colours of high PSI.}
}
\value{
A ggplot2 plot or a pdf.
}
\description{
Nice plot to check expression and PSI across ENCODE mouse developmental stages of different tissues.
}
\examples{
# Save plot of Pax6 expression and its exon 6 PSI 
get_mouse_tissue_devel_tbl(vst_id = "MmuEX0033804", ensembl_gene_id = "ENSMUSG00000027168")  |>
         plot_mouse_tissue_devel(legend = "outside", save_plot = T)
}
