% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprofile_utils.R
\name{tidy_hPTMs}
\alias{tidy_hPTMs}
\title{Parse the information about histone PTMs from an EpiProfile output import}
\usage{
tidy_hPTMs(data)
}
\arguments{
\item{data}{A tibble imported with \code{read_EpiProfile_histone_ratio()}.}
}
\value{
A tidy tibble
}
\description{
Parse the information about histone PTMs from an EpiProfile output import
}
\examples{
read_EpiProfile_histone_ratio("path/to/histone_ratio.xls") |>
    tidy_hPTMs()   
}
